Adding county based covid data test
