+++ stats/src/lib.rs
@@ -1,3 +1,6 @@
+#![feature(test)]
+extern crate test;
+
 pub mod lisa;
