+++ stats/src/lisa.rs
@@ -1,6 +1,3 @@
-//use std::async_iter::from_iter;
-
-
 use geo::GeoFloat;
 use geo_weights::weights::{Weights,TransformType, self};
 use nalgebra::DVector;
@@ -25,7 +22,7 @@ pub enum Quad{
     LL
 }
 
-pub fn lisa (weights: &Weights,values: &[f64], permutations: usize)-> LISAResult 
+pub fn lisa (weights: &Weights,values: &[f64], permutations: usize, keep_sims: bool)-> LISAResult 
 {
     let x = DVector::from_column_slice(values);
 
@@ -38,11 +35,14 @@ pub fn lisa (weights: &Weights,values: &[f64], permutations: usize)-> LISAResult
 
     let quads : Vec<Quad> = x_z.iter().zip(lags.iter()).map(|(a,b)|
         match (a,b){
-            (a,b) if *a >0.0 && *b>0.0 => Quad::HH,
-            (a,b) if *a >0.0 && *b<0.0 => Quad::HL,
-            (a,b) if *a <0.0 && *b>0.0 => Quad::LH,
+            (a,b) if *a >=0.0 && *b >= 0.0 => Quad::HH,
+            (a,b) if *a >=0.0 && *b<0.0 => Quad::HL,
+            (a,b) if *a <0.0 && *b>=0.0 => Quad::LH,
             (a,b) if *a <0.0 && *b<0.0 => Quad::LL,
-            (_,_) => unreachable!()
+            (_,_) => {
+                println!("Got to edge case {}, {}", a,b);
+                unreachable!()
+            }
 
         }
     ).collect();
@@ -85,7 +85,9 @@ pub fn lisa (weights: &Weights,values: &[f64], permutations: usize)-> LISAResult
         if permutations - larger < larger{
             larger = permutations - larger  
         }
-        sims.push(sim_vals);
+        if(keep_sims){
+            sims.push(sim_vals);
+        }
 
         p_vals.push( (larger as f64 +1.0 )/ (permutations as f64+1.0));
 
@@ -101,4 +103,59 @@ pub fn lisa (weights: &Weights,values: &[f64], permutations: usize)-> LISAResult
 }
 
 
+#[cfg(test)]
+mod tests{
+    use super::*;
+    use test::Bencher;
+
+    use geo_weights::{QueensWeights, WeightBuilder};
+    use geo::GeometryCollection;
+    use geojson::{quick_collection, GeoJson};
+
+
+    #[bench]
+    fn real_data_small(b: &mut Bencher){
+        let jsonfile = std::fs::read_to_string( format!("{}/{}",std::env::var("CARGO_MANIFEST_DIR").unwrap(),"test_data/guerry.geojson" )).unwrap();
+        let geojson: GeoJson = jsonfile.parse().unwrap();
+        let geoms: GeometryCollection<f64> = quick_collection(&geojson).unwrap();
+        let weight_builder = QueensWeights::new(10000.0);
+        let weights = weight_builder.compute_weights(&geoms.0);
+        
+        if let GeoJson::FeatureCollection(fc) = geojson{
+            let values: Vec<f64> = fc.features.iter().map(|f| f.property("Donatns").unwrap().as_f64().unwrap()).collect();
+                
+            b.iter(||{
+                lisa(&weights,&values,9999, false);
+            })
+
+        }
+        else{
+            panic!("Expected data to be a feature collection")
+        }
+
+    }
+
+    #[bench]
+    fn real_data_large(b: &mut Bencher){
+        let jsonfile = std::fs::read_to_string( format!("{}/{}",std::env::var("CARGO_MANIFEST_DIR").unwrap(),"test_data/covid.geojson" )).unwrap();
+        let geojson: GeoJson = jsonfile.parse().unwrap();
+        let geoms: GeometryCollection<f64> = quick_collection(&geojson).unwrap();
+        let weight_builder = QueensWeights::new(10000.0);
+        let weights = weight_builder.compute_weights(&geoms.0);
+        
+        if let GeoJson::FeatureCollection(fc) = geojson{
+            let values: Vec<f64> = fc.features.iter().map(|f| f.property("cases").unwrap().as_f64().unwrap()).collect();
+                
+            b.iter(||{
+                lisa(&weights,&values,998, false);
+            })
+
+        }
+        else{
+            panic!("Expected data to be a feature collection")
+        }
+
+    }
+
+}
