+++ stats/tests/lisa.rs
@@ -68,7 +68,7 @@ fn real_data(){
         let values: Vec<f64> = fc.features.iter().map(|f| f.property("Donatns").unwrap().as_f64().unwrap()).collect();
 
 
-        let lisa_results = lisa(&weights,&values,9999);
+        let lisa_results = lisa(&weights,&values,9999,true);
         println!("{:#?}, {:#?}", lisa_results.moran_val, lisa_results.p_vals);
         let j = serde_json::to_string(&lisa_results).unwrap();
         let mut file = File::create("results_real.json").unwrap();
@@ -80,24 +80,3 @@ fn real_data(){
 
 }
 
-#[bench]
-fn real_data_bench(b: &mut Bencher){
-    let jsonfile = std::fs::read_to_string( format!("{}/{}",std::env::var("CARGO_MANIFEST_DIR").unwrap(),"test_data/guerry.geojson" )).unwrap();
-    let geojson: GeoJson = jsonfile.parse().unwrap();
-    let geoms: GeometryCollection<f64> = quick_collection(&geojson).unwrap();
-    let weight_builder = QueensWeights::new(10000.0);
-    let weights = weight_builder.compute_weights(&geoms.0);
-    
-    if let GeoJson::FeatureCollection(fc) = geojson{
-        let values: Vec<f64> = fc.features.iter().map(|f| f.property("Donatns").unwrap().as_f64().unwrap()).collect();
-            
-        b.iter(||{
-            black_box(lisa(&weights,&values,9999));
-        })
-
-    }
-    else{
-        panic!("Expected data to be a feature collection")
-    }
-
-}
